+++ admin_view/nuxt-project/pages/mypage.vue
@@ -1,3 +1,22 @@
 <template>
-  <h1>Hello</h1>
+  <div>
+    <Header/>
+      <h1>Hello</h1>
+  </div>
 </template>
+
+<script>
+import Header from '~/components/Header.vue'
+
+export default {
+  components: {
+    Header,
+  },
+  data() {
+    return {
+    }
+  },
+  methods: {
+  }
+}
+</script>
