+++ view/vue-project/src/views/MyPage.vue
@@ -32,7 +32,7 @@ export default {
     News,
     Regist,
     UserNav,
-　},
+  },
   data () {
     return {
       data: [
